SFU 미디어 서버에서 코드 컴파일 기능 재구현
