+++ media-server/package.json
@@ -13,6 +13,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "axios": "^0.26.0",
     "express": "^4.17.1",
     "httpolyglot": "^0.1.2",
     "mediasoup": "^3.8.2",
